requirements.txt
ecdsa
requests
fastapi
pydantic
uvicorn
streamlit
redis
python-dotenv
docker